+++ src/main/kotlin/de/lancom/genesis/java/GenesisJavaExtension.kt
@@ -3,15 +3,9 @@ package de.lancom.genesis.java
 import com.github.spotbugs.snom.SpotBugsExtension
 import com.github.spotbugs.snom.SpotBugsPlugin
 import com.github.spotbugs.snom.SpotBugsTask
-import org.gradle.api.JavaVersion
 import org.gradle.api.Project
 import org.gradle.api.plugins.JavaPluginExtension
 import org.gradle.api.plugins.quality.*
-import org.gradle.api.publish.PublishingExtension
-import org.gradle.api.publish.maven.MavenPublication
-import org.gradle.api.publish.plugins.PublishingPlugin
-import org.gradle.api.tasks.SourceSet
-import org.gradle.api.tasks.SourceSetContainer
 
 open class GenesisJavaExtension(
     private val project: Project
@@ -20,40 +14,61 @@ open class GenesisJavaExtension(
 
     @JvmOverloads
     fun withCheckstyle(
-        checkstyleVersion: String = "8.39"
+        checkstyleVersion: String = "9.2.1"
     ) {
         project.pluginManager.apply(CheckstylePlugin::class.java)
         project.extensions.configure(CheckstyleExtension::class.java) {
             it.toolVersion = checkstyleVersion
+            it.configFile = project.rootProject.file("config/checkstyle/checkstyle.xml")
+        }
+        project.tasks.named("checkstyleTest", Checkstyle::class.java) {
+            it.enabled = false
         }
     }
 
     @JvmOverloads
     fun withPmd(
-        pmdVersion: String = "6.21.0"
+        pmdVersion: String = "6.41.0"
     ) {
         project.pluginManager.apply(PmdPlugin::class.java)
         project.extensions.configure(PmdExtension::class.java) {
             it.toolVersion = pmdVersion
-            val pmdConfig = project.file("config/pmd/pmd.xml")
+            val pmdConfig = project.rootProject.file("config/pmd/pmd.xml")
             if (pmdConfig.exists()) {
                 it.ruleSets = emptyList()
                 it.ruleSetFiles = project.files(pmdConfig)
             }
         }
+        project.tasks.named("pmdTest", Pmd::class.java) {
+            it.enabled = false
+        }
     }
 
     @JvmOverloads
     fun withSpotBugs(
-        spotbugsVersion: String = "4.2.0"
+        spotbugsVersion: String = "4.5.3"
     ) {
         project.pluginManager.apply(SpotBugsPlugin::class.java)
         project.extensions.configure(SpotBugsExtension::class.java) {
             it.toolVersion.set(spotbugsVersion)
-            val spotbugsConfig = project.file("config/spotbugs/spotbugs.xml")
-            if (spotbugsConfig.exists()) {
-                it.includeFilter.set(spotbugsConfig)
+            val includeConfig = project.rootProject.file("config/spotbugs/include.xml")
+            if (includeConfig.exists()) {
+                it.includeFilter.set(includeConfig)
             }
+
+            val excludeConfig = project.rootProject.file("config/spotbugs/exclude.xml")
+            if (excludeConfig.exists()) {
+                it.excludeFilter.set(excludeConfig)
+            }
+        }
+        project.tasks.named("spotbugsMain", SpotBugsTask::class.java) {
+            it.reports.maybeCreate("html").apply {
+                require(true)
+                outputLocation.set(project.layout.buildDirectory.file("reports/spotbugs/main.html"))
+            }
+        }
+        project.tasks.named("spotbugsTest", SpotBugsTask::class.java) {
+            it.enabled = false
         }
     }