+++ src/main/kotlin/de/lancom/genesis/java/GenesisJavaPlugin.kt
@@ -8,6 +8,6 @@ import org.gradle.api.tasks.compile.JavaCompile
 class GenesisJavaPlugin : Plugin<Project> {
     override fun apply(project: Project) {
         project.pluginManager.apply(JavaPlugin::class.java)
-        project.extensions.create("genesisJava", GenesisJavaExtension::class.java, project)
+        project.extensions.create("genesisJava", GenesisJavaExtension::class.java)
     }
 }