+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2022-01-07
+### Added
+- define checkstyle path in build gradle
+- allow specifying include and exclude filters for spotbugs
+
 ## [1.0.0] - 2020-01-20
 ### Added
 - First release of the Java Gradle plugin